device refactor: move generate and extract_samples into ForestDevice
